fix: move orchestra config to config layer to resolve circular import
- Move OrchestraConfig and related classes from models to config layer
- Convert models/orchestra_config.py to pure re-export module for backwards compatibility
- Fix architecture violation: config should not depend on models
- Update imports in settings.py and orchestra_settings.py

Resolves #497
